Merge main into the 13-player-object branch (#21)
* added GkScene to build (#12)

checked the checkbox so that gkscene is included in the build

* 14 card data objects (#18)

* created CardData scriptable object

Made a scriptable object that holds the data for a card. Contains `string name`, `string[4] actionNames`, `string[4] actionDescriptions`, and `Sprite art`.

* Tested use of CardData objects

Made some additions to test CardData:
- Added a ToString method
- Made a CardDisplay script that takes a CardData and logs its values on Start
- Created a sample card and passed it in to a CardDisplay to print its value

* added more card attributes

added more attributes that are needed to display the card: origin, title, element, flavor text, and the colors that vary across cards

* 4-add board pdf (#16)

* 7 start menu (#8) (#9)

* designed new start menu scene

Created a new start menu scene with a background image and a panel with the title and 2 buttons. Hooked the buttons up to the Canvas's StartMenuOnclick script.

* made buttons functional

made the buttons change scene/quit, changed some build settings to allow us to change scene during runtime

Co-authored-by: Gunnar Dahl <73252935+gadahl@users.noreply.github.com>

* Imported board png into GkScene

Added board into the scene, and png file to assets folder.

* Add GKScene to build settings

* Revert "Add GKScene to build settings"

This reverts commit c17669d7b25ef14c09ccde5db4a23a55fd606cca.

* Add GKScene to build settings

* Merge main to add-board-pdf (#17)

* 7 start menu (#8)

* designed new start menu scene

Created a new start menu scene with a background image and a panel with the title and 2 buttons. Hooked the buttons up to the Canvas's StartMenuOnclick script.

* made buttons functional

made the buttons change scene/quit, changed some build settings to allow us to change scene during runtime

* added GkScene to build (#12)

checked the checkbox so that gkscene is included in the build

---------

Co-authored-by: Gunnar Dahl <73252935+gadahl@users.noreply.github.com>

* Add canvas object and buttons for blue gate

Note: the camera is not focused on the canvas so nothing is being displayed when GKScene is run.

* Converted renderer to image

Converted the game board SpriteRenderer to a UI Image so that it displays on the UI properly. Also made the Image object the parent of the buttons.

---------

Co-authored-by: Julian <43455717+Bluju@users.noreply.github.com>
Co-authored-by: Gunnar Dahl <73252935+gadahl@users.noreply.github.com>
Co-authored-by: Julian <sahaguj@pnw.edu>
Co-authored-by: Gunnar Dahl <gunnardahl4@gmail.com>

---------

Co-authored-by: Gunnar Dahl <73252935+gadahl@users.noreply.github.com>
Co-authored-by: mcdowellbusiness <130783142+mcdowellbusiness@users.noreply.github.com>
Co-authored-by: Gunnar Dahl <gunnardahl4@gmail.com>