41 dice roll animation (#42)
* Added particle effect

Changed the canvas to camera space so that the particle system can appear over it. Also added a canvas to the magnifier so that it appears over the particle system.

* Fixed layers and made basic dice roll

Figured out how to put Canvas elements and Sprites in front of/behind each other while maintaining mouse click functionality, made a dice object that has an ugly dice roll animation on pointer down

* Better dice animation, bouncy walls on release

Gave the dice a better animation so that it spins, shakes, and changes face when the mouse is held down. On release it slows down and stops after a set amount of time. When the dice is released, it is teleported in-bounds and some boundary walls turn on.

* reorganized DiceRoll script

Added comments to make it more clear how the dice roll works, factored out functions `UpdateHeldDice` and `UpdateSlidingDice`

* organization before pull request

* Remove temp debug message

---------

Co-authored-by: Julian <sahaguj@pnw.edu>