28 character select menu (#45)
* Create CharSelectScene

Created the CharSelectScene and CharSelectMenuOnclick Script. The character select menu now has a 'start game' button that will change the scene to GkScene and an arrow button that changes the scene to StartScene.

* Resized canvases
Resized StartScene and CharSelectScene canvases, also resized elements on them

* Add all cards to CharSelectScene

All of the cards that can be selected are now displayed on the CharSelectScene.

* Change Button Sizing & Create functions for buttons

* Add Random Card

The Random card will randomly assign an available card to the player when the "start game" button is clicked.

* Add Player 2 to CharSelectScene

Add visual placeholders for Player 2.

* add @mcdowellbusiness GameManager Script

* Add persistent data between scenes

Set up data persistence between CharSelectScene and GKScene using tutorial:
https://learn.unity.com/tutorial/implement-data-persistence-between-scenes#634f8281edbc2a65c86270c5

* Add AddPlayerButton and update PlayerList

Made PlayerInfo a ScriptableObject, added an AddPlayer2Button that adds Player2 Object to PlayerList  and updates the total amount of players.

* Make AddPlayerButton Singleton

AddPlayerButton will no longer add multiple instances of the same player after clicking multiple times

* Add functionality to RemovePlayerButton

Made the "Remove Player" button beneath Player 2 remove Player2 from the list of players

* Move ui elements

Experimenting with a different way to display players to more effectively utilize screen space.

* did some cleanup

deleted an extra button, changed canvas to camera screen space

* Finished the basic PlayerInfo

* Add all Players and Finish CharSelectScene

Added all of the Scriptable objects for the 4 players and added all of the ui elements for each of the players. Note that Player 1 does not have an add/remove button because we currently want there to be at least 1 player.

* Remove unneeded comment

* Add the dice back to GKScene

I somehow didn't notice the dice wasn't in GKScene anymore

* Final touches

Removed extra dice and put correct PlayerInfos in CardQueue

---------

Co-authored-by: Gunnar Dahl <gunnardahl4@gmail.com>