49 connect information between objects (#51)
* Create GameManager, Add Stubs

* Moved player & gate methods

Implemented PlayerAttacksPlayer, PlayerChangeHealth, and GateChangeHealth in GameManager. Removed the corresponding methods in the other classes. Reformatted GateClasses, PlayerInfo, and TraitHandler.

* Moved & removed a lot of assets

Moved the hype_board image into the image folder, re-made the Battling SO, re-deleted PlyerActions, removed the TextMeshPro/Examples folder, made the Dice Barrier into a prefab, made the StartScene use screen space: camera

* Made debugging TraitHandler easier

Made TraitHandler a static class and added a TraitDebugger object & script. Also made PlayerInfo reset when it is loaded into a scene.

* Reorganized SO's and factored out "Cards" folder

* Created PlayerListSO & implemented it

Created a scriptable object that simply holds a list of PlayerSO's. Used a PlayerListSO to link the CardQueue and the MainManager together so that now they use the same list. Also commented out some of CardQueue's methods.

* Add more traits

* Remove traitHandler Start()

Removed the Start() method in TraitHandler, added a warning for Trait.allMinus1HP since there may be issues with the current implementation.

---------

Co-authored-by: Gunnar Dahl <gunnardahl4@gmail.com>