58 Make classes more accessible (#59)
* Made most vanilla classes Serializable

Made the Gate, StateMachine, and States Serializable. This means if a monobehavior has one of these as a field, you can see inside of it and edit its values.

* Converted TraitHandler to SO

Converted TraitHandler to an SO so that it can be easily plugged into serialized fields. Fixed the classes that depend on TraitHandler . Also deleted the GameStateManager prefab because that script doesn't exist anymore.