71 clean and refactor code (#72)
* Remove Ghost Code

Removed commented out sections of code. If we need these at some point in the future, we can look back at previous versions of the code base.

* Made StateMachine a MonoBehaviour

Removed StateTester and made StateMachine a MonoBehaviour. Reverted all the weird workarounds that existed to get the StateMachine reference to different classes.

* Removed DiceRoll's endEvent

We weren't using it anymore

* Made TraitHandler a MonoBehaviour

Because there was no point for it to be a ScriptableObject

* Made Gate a ScriptableObject

Decided it would be easiest to work with the Gates if they were represented with ScriptableObjects.

* Put some scripts into folders

Made 3 new folders for scripts: Button Scripts, Debug Scripts, and Purely-Visual Scripts. Hopefully this will help new contributors find their way around.

* Updated a few comments

* Rename Folders to be PascalCase

* Rename Prefabs with PascalCase

---------

Co-authored-by: bluju <43455717+Bluju@users.noreply.github.com>
Co-authored-by: Julian <sahaguj@pnw.edu>