73 dealing damage to gates (#74)
* Modify NextTurn in GameManager

NextTurn now shifts the player list but an index out of bounds is called during the second player's trait roll.

* Filled trait list for each card

Made all the cards other than Larry use +1 Health as their trait for 1-4.
Also added a console warning if a card does not have 4 traits

* Added BreakingGateState

Made a BreakingGateState for the game to go into when the player is about to break the gate. Also cleaned up some temp code.

* Moved FinishRollWithValue

Took FinishRollWithValue out of DiceRoll and put it in GameManager, where it is called "UseRollResult."

Gave DiceRoll an event so that it can let GameManager know when to call UseRollResult.

* Pulled state transition out of GateChangeHealth

We want `TransitionTo` to only happen right at the end of an event handler function or a built-in unity function. So I moved the transition out of GateChangeHealth and made it return a bool instead.

* Made UseRollResult do the break effect

UseRollResult now does the dice break effect and heals the gate back to its starting health.

* Remove comment

Yeah I meant it shouldn't be in DiceRoll, it's good now.

---------

Co-authored-by: Julian <sahaguj@pnw.edu>
Co-authored-by: bluju <43455717+Bluju@users.noreply.github.com>