83 implement more traits (#101)
* Update Comments

* Add Stockade Checks to Traits

* Skip Turn on TraitRoll = 6

* Add Stockade Checks to Traits

* Implement Trait reduceGateDamage

* Clarify Trait Inspector Names

* Move Stockade Checks to GameManager

Stockade checks are now in PlayerAttacksPlayer() and PlayerChangeHealth() functions of GameManager instead of in TraitHandler.cs

* Implement Trait noDamageTurn and Bugfix

Implemented noDamageTurn where a player will no longer take damage on their turn if they activate this trait. Fixed a bug that healed 1 health instead of dealing 1 damage to other players on Trait.allMinus1HP.

* Fix noDamageTurn and Add resetEffects

Accidentally made noDamageTurn disable after blocking damage once, now it will be disabled at the end of the player's turn. Added resetEffects function in PlayerSO to reset all boolean variables we are using for temporary trait effects.

* Fix Dead Current Player Bug

If a player died during their turn due to their trait roll using PlayerChangeHealth they would continue their turn even if dead. Their turn will now end when they die.

* Implement doubleGateAbil

* Implement increaseGateDamage & rename minus2gate

The trait 'minus2gate' is meant to deal 2 damage to the gatekeeper. This name was ambiguous so I changed it to minus2GateKeeper for clarity.

* Add deal3Dam Trait

* fixed lots of stuff

- TraitHandler.ActivateTrait now returns a trait to be transitioned to outside
- fixed green gate
- Gave TraitHandler a dedicated Hide method
- removed asserts from OnPointerExit, -Enter, and -Click
- CardMagnifier is no longer initially active in CharSelectScene
- I no longer trust OnEnable so I rewrote some scripts to use Initialize methods instead

* Rework Multiplier Traits

Converted many trait effects that work as damage multipliers to integers instead of bools. Now instead of checking if the trait is active, we multiply the value of the trait in a formula to calculate damage.

* Add deal2Dam Trait

* Remove Dead Players from PlayerSelection

* Give cards correct traits

---------

Co-authored-by: Gunnar Dahl <gunnardahl4@gmail.com>