107 store game stats (#122)
* Transition Scene if Player Wins by Survival

Made a CheckWinBySurvival() function that checks if there is only one player remaining. If there is only one player remaining, then that player is saved in Globals.winningPlayer and then the scene is transitioned to EndScene. Some GameManager functions needed to be changed to non-static as well.

* Transition Scene when player kills Gatekeeper

* Add stats variables

* Reset player health and stats on new game

Use PlayerSO's GameReset() to reset players at the start of a new game
Close Issue #120

* Transition to next turn if current player dies in battle

* init stats variables and add documentation

* Update stats variables during game

* update damage to other players and gatekeeper

* Increment totalStockade in all locations

* move player reset to StartGame