clean up previous state
this introduces a new method to "manually" switch from one state to another and uses it.  this is due to the current UI state now being a static variable.  this might be a dirty way to do it, but it seems like the only way.
TODO:  IP and player list is not networked