123 add stats to endgamescene (#180)
* Clean scene

* Display Player Card on CardDisplay

* Display Winner Stats

* Add Placeholders for 4 Players

Note that this does not currently adjust to the amount of players in the game and will throw errors with less than 4 players.

* Make Card Displays Match Player List

A display will be made for each player in the player list

* Create EndStatDisplay Prefab

* Create displays and add players to them

This still doesn't display the statDisplays on the scene.

* Displays correctly now

* Update EndSceneInfo.cs

* Add stats to end screen display

* Add Winner Text and Change Spacing

* Change EndScene to ResultScene

Renamed EndScene and cleaned up the EndSceneInfo script.

* Update ResultScene UI and font asset

Fixed Main Menu Button and added the CursorPreference to it so the cursor is a pointer when hovering over it.

---------

Co-authored-by: Gunnar Dahl <73252935+gadahl@users.noreply.github.com>